+++ JobModule/jobmodule_example_2sshconnection.py
@@ -4,8 +4,10 @@ import select
 import time
 import LoggingMgr
 from pprint import pprint
-import jobfrag_sshconn
-import jobmodule_base
+#import jobfrag_sshconn
+#import jobmodule_base
+import JobModule.jobfrag_sshconn as jobfrag_sshconn
+import JobModule.jobmodule_base  as jobmodule_base 
 
 DEBUG_MODE = True
